refactor: extract filtres statuts
